Add files via upload
Please, keep in mind that this Plugin has not been released to the public and it's only used in one personal blog. I works ok, but it's a long time since it was updated. A new version will be released in Nov 2017. Feel free to test it, but it's not recommnended to use it in production.